pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bubble-analysis"
version = "0.3.0"
description = "Static analysis tool for tracing exception flow through Python codebases"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    {name = "Ian McLaughlin", email = "ianm199@github.com"},
]
keywords = [
    "static-analysis",
    "exceptions",
    "python",
    "code-analysis",
    "flask",
    "fastapi",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Typing :: Typed",
]
dependencies = [
    "libcst>=1.1.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "msgpack>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/ianm199/flow"
Repository = "https://github.com/ianm199/flow"
Issues = "https://github.com/ianm199/flow/issues"

[project.scripts]
bubble = "bubble.cli:app"

[tool.setuptools.packages.find]
include = ["bubble*"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "strict"
exclude = ["examples"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "E402"]

[tool.ruff.lint.isort]
known-first-party = ["bubble"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
